% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-qview.R
\name{is_qview}
\alias{is_qview}
\title{Test whether an object is a \code{qview}}
\usage{
is_qview(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
Logical scalar.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
path <- system.file("extdata", "example.Q-View", package = "qviewparsR")
is_qview(read_qview(path, verbose = FALSE))
is_qview(list())

}
\seealso{
Other qview-methods:
\code{\link[=as_tibble.qview]{as_tibble.qview()}},
\code{\link[=plot.qview]{plot.qview()}},
\code{\link[=print.qview]{print.qview()}},
\code{\link[=summary.qview]{summary.qview()}}
}
\concept{qview-methods}
