% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-qview.R
\name{print.qview_summary}
\alias{print.qview_summary}
\title{Print a qview_summary object}
\usage{
\method{print}{qview_summary}(x, ...)
}
\arguments{
\item{x}{A \code{qview_summary} object returned by \code{\link[=summary.qview]{summary.qview()}}.}

\item{...}{Unused; for S3 generic compatibility.}
}
\value{
\code{x}, invisibly.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
path <- system.file("extdata", "example.Q-View", package = "qviewparsR")
qv <- read_qview(path, verbose = FALSE)
summary(qv)

}
\seealso{
Other qview-methods:
\code{\link[=as_tibble.qview]{as_tibble.qview()}},
\code{\link[=is_qview]{is_qview()}},
\code{\link[=plot.qview]{plot.qview()}},
\code{\link[=print.qview]{print.qview()}},
\code{\link[=summary.qview]{summary.qview()}}
}
\concept{qview-methods}
