% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_qview.R
\name{qview_to_csv_dir}
\alias{qview_to_csv_dir}
\title{Write a Q-View object as a directory of CSV files}
\usage{
qview_to_csv_dir(qv, dir, template = NULL, call = rlang::caller_env())
}
\arguments{
\item{qv}{A \code{qview} object from \code{\link[=read_qview]{read_qview()}}.}

\item{dir}{Output directory (created if it does not exist).}

\item{template}{Optional plate-template tibble to include.}

\item{call}{The execution environment of the calling function. Used
for error reporting; experts only.}
}
\value{
\code{qv}, invisibly, to support pipelines.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\seealso{
Other qview-export: 
\code{\link{qview_to_xlsx}()}
}
\concept{qview-export}
