% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qviewparsR-package.R
\docType{package}
\name{qviewparsR-package}
\alias{qviewparsR}
\alias{qviewparsR-package}
\title{qviewparsR: Read .Q-View Multiplex ELISA Project Files}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Pure-R parser for \code{.Q-View} binary project files used in
chemiluminescent multiplex ELISA plate imaging and quantification.
Extracts pixel intensities, analyte mappings, sample assignments,
plate layout, optional curve-fit parameters, and the embedded CSV
report from the H2 database container format. Returns tidy tibbles
ready for downstream statistical analysis.
}
\section{Main functions}{

\itemize{
\item Reader: \code{\link[=read_qview]{read_qview()}} -- parse a \code{.Q-View} file.
\item Helpers: \code{\link[=strip_qview_prefix]{strip_qview_prefix()}} reverse the Q-View internal naming
convention; \code{\link[=well_label]{well_label()}} map (row, column) to plate notation.
\item Optional: \code{\link[=read_qview_template]{read_qview_template()}} parse a Q-View well-assignment
template CSV.
\item Methods: \code{\link[=is_qview]{is_qview()}}, \code{\link[=print.qview]{print.qview()}}, \code{\link[=plot.qview]{plot.qview()}},
\code{\link[=as_tibble.qview]{as_tibble.qview()}}.
\item Shiny: \code{\link[=qview_app]{qview_app()}} interactive upload / preview / download.
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/CTTIR/qviewparsR}
  \item \url{https://cttir.github.io/qviewparsR/}
  \item Report bugs at \url{https://github.com/CTTIR/qviewparsR/issues}
}

}
\author{
\strong{Maintainer}: Raban Heller \email{raban.heller@charite.de} (\href{https://orcid.org/0000-0001-8006-9742}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Raban Heller \email{raban.heller@charite.de} (\href{https://orcid.org/0000-0001-8006-9742}{ORCID}) [copyright holder]
  \item Marco Mannes
}

}
\keyword{internal}
